Revert "MemeUnitBot pre-prod"
This reverts commit e7ba4cbb45a6e290c4db45a6a6cd6cb6a32d6aaf.
