+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.MemeUnit</groupId>
-    <artifactId>MemeUnitBot</artifactId>
+    <groupId>org.MINDN</groupId>
+    <artifactId>MindBlownBot</artifactId>
     <version>1.0.0</version>
 
     <properties>
